requirements.txt
fastapi
uvicorn
streamlit
pypdf
faiss-cpu
openai
reportlab
python-dotenv
